backend/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
httpx
